1. added map node capability to apply warp and apply transform, and added relevant tests, 2. consolidated mock code from tests into a seperate 'mocks' module
